Replace Svelte app with Sveltekit, update existing examples
