--- svelte-ids-wc/.nvmrc
@@ -1,1 +1,0 @@
-16.15.1
