--- svelte-ids-wc/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "svelte-ids-wc",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "rollup": "^2.75.7",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-sourcemaps": "^0.6.3",
-    "rollup-plugin-svelte": "^7.1.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "sirv-cli": "^2.0.2",
-    "svelte": "^3.48.0"
-  },
-  "dependencies": {
-    "ids-enterprise-wc": "1.0.0-beta.0"
-  }
-}
