--- svelte-ids-wc/src/App.svelte
@@ -1,13 +1,0 @@
-<script>
-	import { default as IdsTagStaticExample } from "./examples/ids-tag/Index.svelte";
-	import { default as IdsTagDynamicExample } from "./examples/ids-tag/Dynamic.svelte";
-	import { default as IdsButtonIndexExample } from './examples/ids-button/Index.svelte';
-	import { default as IdsModalIndexExample } from './examples/ids-modal/Index.svelte';
-</script>
-
-<style></style>
-
-<IdsTagStaticExample />
-<IdsTagDynamicExample />
-<IdsButtonIndexExample />
-<IdsModalIndexExample />
