--- svelte-ids-wc/src/main.js
@@ -1,10 +1,0 @@
-import App from './App.svelte';
-
-const app = new App({
-	target: document.body,
-	props: {
-		color: '#f00'
-	}
-});
-
-export default app;
