+++ sveltekit-ids-wc/doc/CHECKLIST.md
@@ -1,0 +1,115 @@
+# Completion Checklist
+
+## Components To Convert
+
+Components marked completed in this checklist will have both an example of direct usage AND an example using a Svelte component wrapper
+
+ - [ ] About (ids-about)
+ - [ ] Accordion (ids-accordion)
+ - [ ] Action Sheet (ids-action-sheet)
+ - [ ] Alerts (ids-alert)
+ - [ ] Application menu (ids-application-menu)
+ - [ ] Area (ids-area-chart)
+ - [ ] Arrange (ids-swapabble)
+ - [ ] Autocomplete (ids-autocomplete)
+ - [ ] Badges (ids-badge)
+ - [ ] Bar (ids-bar-chart)
+ - [ ] Block grid (ids-block-grid)
+ - [ ] Breadcrumb (ids-bread-crumb)
+ - [ ] Bubble (ids-bubble-chart)
+ - [ ] Button (ids-button)
+ - [ ] Bullet (ids-bullet-chart) skipping until needed
+ - [ ] Builder (ids-patterns) skipping until needed
+ - [ ] Button (ids-button)
+ - [ ] Calendar (ids-calendar)
+ - [ ] Cards (ids-card)
+ - [x] Checkboxes (ids-checkbox)
+ - [ ] Circle Pager (ids-scroll-view)
+ - [ ] Color Picker (ids-color-picker)
+ - [ ] Column (ids-column-chart or ids-bar with a orientation setting)
+ - [ ] Completion Chart (ids-progress-chart)
+ - [ ] Contextual Action Panel (ids-contextual-action-panel)
+ - [ ] Datagrid (ids-data-grid)
+ - [ ] Data Label (ids-data-label)
+ - [ ] Datepicker (ids-date-picker)
+ - [ ] Donut (ids-donut or combined with ids-pie)
+ - [ ] Drag (ids-draggable)
+ - [ ] Dropdown (ids-drop-down)
+ - [ ] Editor (ids-editor)
+ - [ ] Emptymessage (ids-empty-message)
+ - [ ] Error Page (ids-page-error)
+ - [ ] Expandable area (ids-expandable-area)
+ - [ ] Fieldset (ids-field-set)
+ - [ ] Fontpicker (added as part of editor)
+ - [ ] Form  (ids-form)
+ - [ ] Grid (ids-layout-grid)
+ - [ ] Header (ids-header)
+ - [ ] Hierarchy (ids-hierarchy)
+ - [ ] Homepage (ids-homepage)
+ - [ ] Hyperlinks (ids-hyperlink)
+ - [ ] Icons (ids-icon)
+ - [ ] Images (ids-image)
+ - [ ] Input (ids-input)
+ - [ ] (Layout) Responsive Grid (ids-layout-grid)
+ - [ ] Line (ids-line-chart)
+ - [ ] Listbuilder (ids-list-builder)
+ - [ ] Listview (ids-list-view)
+ - [ ] Loading indicator (ids-loading-indicator) aka Busy Indicator / Loader
+ - [ ] Locale (ids-locale)
+ - [ ] Lookup (ids-lookup)
+ - [ ] Mask (ids-mask)
+ - [ ] Masthead (ids-masthead)
+ - [ ] Menu (ids-menu)
+ - [ ] MenuButton (ids-menu-button)
+ - [ ] Message (ids-message)
+ - [ ] Modal (ids-modal)
+ - [ ] Monthview (ids-month-view)
+ - [ ] Multiselect (ids-multi-select)
+ - [ ] Notification (ids-notification-banner)
+ - [ ] Page Layouts (ids-page-layout or as individual separate e amples)
+ - [ ] Page Patterns (ids-page-pattern or as individual separate e amples)
+ - [ ] Pager (ids-pager)
+ - [ ] Personalize (ids-personalize or as amixin on components thats support)
+ - [ ] Pie (ids-pie-chart)
+ - [ ] Popup (ids-popup)
+ - [ ] Popupmenu (ids-men
+ - [ ] Positive Negative (ids-positive-negative-chart) skipping until needed
+ - [ ] Radar (ids-radar-chart) skipping until needed
+ - [ ] Radios (ids-radio)
+ - [ ] Rating (ids-rating)
+ - [ ] Renderloop (mixin)
+ - [ ] Scatterplot (ids-scatter-plot-chart)
+ - [ ] Searchfield (ids-search-field)
+ - [ ] Sign-in (ids-layouts)
+ - [ ] Skiplink (ids-skip-link)
+ - [ ] Slider (ids-slider)
+ - [ ] Sparkline (ids-spark-line-chart) skipping until needed
+ - [ ] Spinbox (ids-spin-box)
+ - [ ] Splitter (ids-splitter)
+ - [ ] Stepchart (ids-step-chart)
+ - [ ] Swaplist (ids-swap-list)
+ - [ ] Switch (ids-switch)
+ - [ ] Tabs (ids-tabs)
+ - [ ] Tabs Header (ids-tabs with option)
+ - [ ] Tabs Module (ids-tabs with option)
+ - [ ] Tabs Multi (ids-tabs with option) skipping until needed
+ - [ ] Tabs Vertical (ids-tabs with option)
+ - [x] Tag (ids-tag)
+ - [ ] Targeted Achievement (ids-progress-chart)
+ - [ ] Textarea (ids-textarea)
+ - [ ] Timeline (ids-time-line) skipping until needed (also have process indicator)
+ - [ ] Timepicker (ids-time-picker)
+ - [ ] Toast (ids-toast)
+ - [ ] Toolbar (ids-toolbar)
+ - [ ] Tooltip (ids-tooltip
+ - [ ] TrackDirty (mixin)
+ - [ ] Tree (ids-tree)
+ - [ ] Treemap (ids-tree-map)
+ - [ ] Trigger Field (ids-trigger-field)
+ - [ ] Typography (ids-text)
+ - [ ] Upload (ids-upload)
+ - [ ] Upload Advanced (ids-upload-advanced)
+ - [ ] Validation (ids-validation-mixin)
+ - [ ] Visibility (ids-hidden)
+ - [ ] Week View (ids-week-view)
+ - [ ] Wizard (ids-wizard)
