+++ sveltekit-ids-wc/src/routes/ids-checkbox/index.svelte
@@ -1,0 +1,9 @@
+<ids-layout-grid auto="true">
+    <ids-text font-size="20" type="h1">Ids Checkbox Examples</ids-text>
+</ids-layout-grid>
+
+<ids-layout-grid auto="true">
+    <ids-layout-grid-cell>
+        <ids-demo-listing label="Checkbox Examples" component-name="ids-checkbox"></ids-demo-listing>
+    </ids-layout-grid-cell>
+</ids-layout-grid>
